Move theme status to context
