+++ app/tailwind.config.ts
@@ -4,7 +4,7 @@ import daisyui from "daisyui";
 import type { Config as DaisyUIConfig } from "daisyui";
 import type { Config } from "tailwindcss";
 
-import { DARK_THEME, LIGHT_THEME } from "./client/utils/themes";
+import { DARK_THEME, LIGHT_THEME } from "./client/utils/themes/themes";
 
 export default {
   content: ["./index.html", "./client/**/*.{js,ts,jsx,tsx}"],
